msm: display: update histogram interface for MDP5
MDP5 supports 256 histogram bins now, which is not able to be
stored in the 8 bit unsigned integer used currently. Increase
this variable to be able to accommodate it.

+++ include/linux/msm_mdp.h
@@ -399,7 +399,7 @@ struct mdp_histogram_start_req {
 	uint32_t block;
 	uint8_t frame_cnt;
 	uint8_t bit_mask;
-	uint8_t num_bins;
+	uint16_t num_bins;
 };
 
 /*
